\name{as.mapping}
\alias{as.mapping}
\title{Coerce correspondance table into a mapping}
\usage{
  as.mapping(data, from, to)
}
\arguments{
  \item{data}{a data frame}

  \item{from}{column of \code{data} that has to be mapped.}

  \item{to}{column of \code{data} that contains the end
  result.}
}
\description{
  Coerce correspondance table into a mapping
}

