\name{find.in.mapping}
\alias{find.in.mapping}
\title{Find the origin value in a mapping}
\usage{
  find.in.mapping(mapping, find)
}
\arguments{
  \item{mapping}{See \code{\link{mapping}}}

  \item{find}{end value of the mapping.}
}
\value{
  the origin value corresponding to \code{find} in
  \code{mapping}.
}
\description{
  Find the origin value in a mapping
}

